backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
sqlalchemy==1.4.53
alembic==1.12.1
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
requests>=2.31.0
redis>=4.5.0
celery>=5.3.0
websockets>=11.0
httpx>=0.25.0
cryptography>=41.0.0
# PostgreSQL support
psycopg2-binary>=2.9.0
# YAML configuration support
PyYAML>=6.0.0
